Switch to Hyper 0.11, the asynchronous version.
I decided that this was probably going to be important to do since I can
envision several use cases where it would be really nice for blobman to act as
an HTTP server, and in that case it really needs to do fancy asynchronous I/O
and all that. So I experimented a bit and I think I figured out how to deal
with hyper's asynchronous Body stream in a synchronous context.